added by mohini nikhare

+++ Agriculture_Equipment_Loan_Application/src/main/java/com/financeCompany/agri/project/appDto/CMCusRespEMIDto.java
@@ -13,7 +13,7 @@ public class CMCusRespEMIDto {
 	private double interest;
 	private double totalAmount;
 	private String rateOfInterest;
-	private double loanTenure;
+	private long loanTenure;
 	private double requiredLoanAmount;
